New file management tab in project settings dialog, fixes: http://www.kdenlive.org/mantis/view.php?id=1140
svn path=/trunk/kdenlive/; revision=3900

+++ src/docclipbase.h
@@ -147,7 +147,7 @@ Q_OBJECT public:
     bool hasFileSize() const;
 
     /** Returns the filesize, or 0 if there is no appropriate filesize. */
-    uint fileSize() const;
+    qulonglong fileSize() const;
 
     /** Returns true if this clip refers to the clip passed in. A clip refers to another clip if
      * it uses it as part of it's own composition. */
